pyproject.toml
[tool.poetry]
name = "ml_experiments"
version = "0.1.0"
description = ""
authors = ["Marian Szabo <mszabo@coralwave.com>"]

[tool.poetry.dependencies]
python = "^3.6"
jupyter = "^1.0.0"
matplotlib = "^3.3.4"
cffi = "^1.14.5"
protobuf = "^3.14.0"
tensorflow-gpu = "^2.4.1"
Keras = "^2.4.3"
#cupy-cuda111 = "^8.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
